Extend cleanbins.cmd to clean folders recursively

+++ git/cleanbins.cmd
@@ -1,10 +1,59 @@
 @echo off
-REM cleanbins.cmd
-REM Deletes all uncommited files (except node_modules and certificates) in the current repo
+REM cleanbins.cmd [/recursive] [path]
+REM Deletes all uncommited files (except node_modules and certificates) in the given git repo
+REM
+REM Options:
+REM
+REM path        The path of the git repo to clean (default: current)
+REM /recursive  Recursively clean each subfolders of the given path
 
 setlocal enabledelayedexpansion
 
-call git clean -f -d -e node_modules/ -e *.pfx -e Package.StoreAssociation.xml -x
+set recursive=0
+
+set target_path=%CD%
+
+:loop
+set part=%1
+if not "%part%"=="" (
+  if "%part%"=="/recursive" (
+      set recursive=1
+  ) else if "%part:~0,1%"=="/" (
+      echo cleanbins.cmd: Unknown flag "%part%"
+      exit /b 1
+  ) else (
+      set target_path=%part%
+  )
+  shift
+  goto :loop
+)
+:loopend
+
+pushd %target_path%
+
+if "%recursive%"=="0" (
+  call :clean
+) else (
+  for /d %%i in (*.*) do (
+    pushd %%i
+    call :clean
+    popd
+  )
+)
+
+popd
+goto :exit
+
+:clean
+if not exist ".git/" (
+  @echo cleanbins.cmd: Path "%cd%" is not a git repo
+) else (
+  @echo cleanbins.cmd: Cleaning "%cd%"
+  call git clean -f -d -e node_modules/ -e *.pfx -e Package.StoreAssociation.xml -x
+)
+exit /b %ERRORLEVEL%
+
+:exit
 
 endlocal
